+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/PeerStateActions.hs
@@ -66,7 +66,7 @@ import Ouroboros.Network.PeerSelection.Governor.Types (PeerStateActions (..))
 import Ouroboros.Network.Protocol.Handshake (HandshakeException)
 import Ouroboros.Network.RethrowPolicy
 
-import Ouroboros.Network.ConnectionHandler (Handle (..), HandleError (..),
+import Ouroboros.Network.ConnectionHandler (Handle (..), HandlerError (..),
            MuxConnectionManager)
 import Ouroboros.Network.ConnectionManager.Types
 import Ouroboros.Network.PeerSelection.PeerSharing (PeerSharing)
@@ -836,29 +836,29 @@ withPeerStateActions PeerStateActionsArguments {
                                    ("peerMonitoringLoop " ++ show remoteAddress))
               pure connHandle
 
-            Right (Disconnected _ Nothing) ->
-              -- Disconnected in 'TerminatingState' or 'TerminatedState' without
-              -- an exception.
-              throwIO $ userError "establishPeerConnection: Disconnected"
-            Right (Disconnected _ (Just reason)) ->
-              case reason of
-                HandleHandshakeClientError err -> do
-                  traceWith spsTracer (PeerStatusChangeFailure
-                                        (ColdToWarm Nothing remotePeerAddr)
-                                        (HandshakeClientFailure err))
-                  throwIO (ClientException err)
+            Right (Disconnected _ disconnectionError) ->
+              case disconnectionError of
+                ConnectionHandlerError handlerError ->
+                  case handlerError of
+                    HandleHandshakeClientError err -> do
+                      traceWith spsTracer (PeerStatusChangeFailure
+                                            (ColdToWarm Nothing remotePeerAddr)
+                                            (HandshakeClientFailure err))
+                      throwIO (ClientException err)
 
-                HandleHandshakeServerError err -> do
-                  traceWith spsTracer (PeerStatusChangeFailure
-                                        (ColdToWarm Nothing remotePeerAddr)
-                                        (HandshakeServerFailure err))
-                  throwIO (ServerException err)
+                    HandleHandshakeServerError err -> do
+                      traceWith spsTracer (PeerStatusChangeFailure
+                                            (ColdToWarm Nothing remotePeerAddr)
+                                            (HandshakeServerFailure err))
+                      throwIO (ServerException err)
 
-                HandleError err -> do
-                  traceWith spsTracer (PeerStatusChangeFailure
-                                        (ColdToWarm Nothing remotePeerAddr )
-                                        (HandleFailure err))
-                  throwIO err
+                    HandlerError err -> do
+                      traceWith spsTracer (PeerStatusChangeFailure
+                                            (ColdToWarm Nothing remotePeerAddr )
+                                            (HandleFailure err))
+                      throwIO err
+
+                _ -> throwIO disconnectionError
       where
         mkAwaitVars :: OuroborosBundle muxMode (ExpandedInitiatorContext peerAddr m)
                                                responderCtx ByteString m a b